LPS-80122 (bookmarks-uad) Add base implementation class for UADAnonymizer
